reorganized import datasets script
